+++ Packages/com.unity.inputsystem/InputSystem/Plugins/Switch/NPad.cs
@@ -331,7 +331,7 @@ namespace UnityEngine.Experimental.Input.Plugins.Switch
     /// An NPad controller for Switch, which can be a Joy-Con.
     /// </summary>
     /// <seealso cref="NPadInputState"/>
-    [InputControlLayout(stateType = typeof(NPadInputState))]
+    [InputControlLayout(stateType = typeof(NPadInputState), displayName = "Switch Controller (on Switch)")]
     public class NPad : Gamepad, INPadRumble
     {
         public ButtonControl leftSL { get; private set; }